// packages/conv-ai-provider/src/types.ts
export enum FeedbackRating {
  Negative = '-1',
  Positive = '1',
}

// expected action schema from ConvAi API
export enum MessageRole {
  User = 'user',
  Assistant = 'assistant',
}

// Every individual message (from user or AI) has a unique ID
export type MessageId = ReturnType<typeof uuidv4>;

// A user prompt message and its resulting AI response message share a request ID
// (This informs us which AI response corresponds to which user prompt)
export type RequestId = ReturnType<typeof uuidv4>;

// schema of actions from websocket notification streaming message chunks
export type WSNotificationBotMessageClientAction = BotMessageClientAction & {
  arguments: {
    text: {
      streamId: string;
    };
  };
};

export enum ConnectionState {
  Connected = 'Connected',
  Reconnecting = `Reconnecting`,
  Closed = `Closed`,
}

export enum MessageRequestType {
  RegenerationRequest = 'RegenerationRequest',
  UserPrompt = 'UserPrompt',
}

type ISODateTimeString = string; // e.g. "2024-05-03T17:32:19.318Z"

export type ConversationRateLimitError =
  | 'UserRateLimitExceeded'
  | 'ServiceRateLimitExceeded'
  | 'TokenLimitExceeded';

export type ConversationError =
  | ConversationRateLimitError
  | 'ContentFilterFailed'
  | 'InternalError';

type MessageEvent = {
  startStreamingResponse: (
    content: string,
    messageId: MessageId,
    requestId: RequestId,
    actions: WSNotificationBotMessageClientAction[],
  ) => void;
  finishStreamingResponse: (requestId: RequestId) => void;
  updateStreamingResponse: (content: string, messageId: MessageId) => void;
  errorStreamingResponse: (conversationError: ConversationError, messageId: MessageId) => void;
  connectionConnected: () => void;
  connectionReconnecting: () => void;
  connectionClose: () => void;
};

export type Emitter = TypedEventEmitter<MessageEvent>;

export type PromptContext = {
  requestType: MessageRequestType;
  prompt: string;
  originalPrompt?: string;
};

export type PromptMessage = {
  role: MessageRole.User;
  requestId?: RequestId;
  messageId?: MessageId;
  content: string;
  context: PromptContext;
  createTime: ISODateTimeString;
};

export type BotMessage = {
  role: MessageRole.Assistant;
  requestId: RequestId;
  messageId: MessageId;
  content: string;
  context: PromptContext;
  createTime: ISODateTimeString;
  actions: WSNotificationBotMessageClientAction[];
};

export type MessageHistoryItem = PromptMessage | BotMessage;

// Conv AI derived types
export type {
  Conversation,
  MessageChunk,
  StreamChunk,
} from '@rbx/client-conv-ai-platform-orchestration/v1';

export enum ConvApiV2ActionType {
  DisplayRichText = 'DisplayRichText',
  DisplayError = 'DisplayError',
}

// ref: https://github.rbx.com/Roblox/grasshopper/blob/master/packages/clients/schemas/convAiPlatformOrchestration/v1.json
export type BotMessageClientAction = {
  actionTime: ISODateTimeString;
  actionSummary: string; // autosummarized by assistant
  actionType: 'client_action';
  arguments: {
    text: string; // assistant message text
    errorId?: string; // error ID if actionType is DisplayError
  };
  command: ConvApiV2ActionType;
  plugin: ''; // empty string for docs assistant
};
// ...
type ConvAiResponse = {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any -- from autogenerated client definition
  [key: string]: any | any;
  conversation: Conversation;
};

export type ConversationState = {
  streaming: boolean;
  isResponseFailed: boolean;
  messageHistory: MessageHistoryItem[]; // messages shown in chat history UI
  errorByMessageHistoryIdx: Map<number, ConversationError>;
  // map of message errors in messageHistory, by messageHistory index
  // - excluded in message history context sent to bot on sendPrompt
  // - shown below message in ui as an alert, per ConversationErrorAlert
  ratingByRequestId: Map<RequestId, FeedbackRating>;
  activePromptMessage?: PromptMessage; // last sent message by user
  activeResponseMessage?: BotMessage; // active response responding to active user prompt
  activeRequestId?: string; // request ID shared with active prompt + active response
  originalPrompt?: string; // if {activePrompt} is based on regenerating a given prompt, store that prompt in {originalPrompt}
  conversationId?: string;
  connectionState: ConnectionState;
  activeMessageConversationError?: ConversationError; // error of activePromptMessage or activeResponseMessage
  userDailyRateLimit?: number;
  userRateLimitErrorCount: number;
};

export enum ConversationLogEvent {
  // Bot events
  BotStartStreamingResponse = 'botStartStreamingResponse',
  BotFinishStreamingResponse = 'botFinishStreamingResponse',
  BotErrorStreamingResponse = 'botErrorStreamingResponse',

  // Connection events
  ConnectionReconnecting = 'connectionReconnecting',
  ConnectionReconnected = 'connectionReconnected',
  ConnectionClosed = 'connectionClosed',

  // User events
  UserSubmitPrompt = 'userSubmitPrompt',
  UserSubmitPromptResponseError = 'userSubmitPromptResponseError',
  UserSubmitRegenerationPrompt = 'userSubmitRegenerationPrompt',
  UserClickStop = 'userClickStop',
  UserRateResponse = 'userRateResponse',
  UserStartsRegenerateMessage = 'userStartsRegenerateMessage',
}

export type AssistantOptIn = {
  accepted: boolean;
  process: (callback: () => void) => void;
};

export type RegenerateResponseProps = {
  originalPrompt: string;
  message: MessageHistoryItem;
  messageLogParams?: Record<string, string>;
  promptAssistantResponse: TPromptAssistantResponse;
};

export type SendUserPromptProps = {
  userPrompt: string;
  examplePromptCategory?: string;
  promptAssistantResponse: TPromptAssistantResponse;
};

export type TPromptAssistantResponse = (state: ConversationState) => Promise<ConvAiResponse>;

export type TLogConversationEvent = (
  eventName: ConversationLogEvent,
  state: ConversationState,
  extraParams?: Record<string, unknown>,
) => void;

export type ConvAiConfig = {
  // The namespace that the conversation notifications are sent to.
  conversationNamespace: string;
  assistantOptIn: AssistantOptIn;
  logConversationEvent: TLogConversationEvent;
};